+++ app/Modules/IncomingMail/Models/IncomingMailFollowUp.php
@@ -34,7 +34,9 @@ class IncomingMailFollowUp extends Model
 
 		static::deleting(function($followup) {
 			/* Remove attachment file*/
-			Upload::delete($followup->path_to_file);
+			if (!empty($followup->path_to_file)) {
+				Upload::delete($followup->path_to_file);
+			}
 		});
     }
 }
