+++ app/Modules/IncomingMail/Models/IncomingMailattachment.php
@@ -21,7 +21,9 @@ class IncomingMailAttachment extends Model
 
 		static::deleting(function($attachment) {
 			/* Remove attachment file*/
-			Upload::delete($attachment->path_to_file);
+			if (!empty($attachment->path_to_file)) {
+				Upload::delete($attachment->path_to_file);
+			}
 		});
     }
 }
